Add support for parameterized remote-store testing for Integration Tests.
