+++ test/framework/src/main/java/org/opensearch/test/ParameterizedStaticSettingsOpenSearchIntegTestCase.java
@@ -11,10 +11,12 @@ package org.opensearch.test;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.indices.replication.common.ReplicationType;
 
+import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
+import static org.opensearch.gateway.remote.RemoteClusterStateService.REMOTE_CLUSTER_STATE_ENABLED_SETTING;
 import static org.opensearch.indices.IndicesService.CLUSTER_REPLICATION_TYPE_SETTING;
 
 /**
@@ -30,6 +32,8 @@ import static org.opensearch.indices.IndicesService.CLUSTER_REPLICATION_TYPE_SET
  */
 public abstract class ParameterizedStaticSettingsOpenSearchIntegTestCase extends ParameterizedOpenSearchIntegTestCase {
 
+    protected static final String REPOSITORY_NAME = "test-remote-store-repo";
+    private Path repositoryPath;
     public static final List<Object[]> replicationSettings = Arrays.asList(
         new Object[] { Settings.builder().put(CLUSTER_REPLICATION_TYPE_SETTING.getKey(), ReplicationType.DOCUMENT).build() },
         new Object[] { Settings.builder().put(CLUSTER_REPLICATION_TYPE_SETTING.getKey(), ReplicationType.SEGMENT).build() }
@@ -39,9 +43,21 @@ public abstract class ParameterizedStaticSettingsOpenSearchIntegTestCase extends
         super(nodeSettings);
     }
 
+    public static final List<Object[]> remoteStoreSettings = Arrays.asList(
+        new Object[] { Settings.builder().put(REMOTE_CLUSTER_STATE_ENABLED_SETTING.getKey(), true).build() },
+        new Object[] { Settings.builder().put(REMOTE_CLUSTER_STATE_ENABLED_SETTING.getKey(), false).build() }
+    );
+
     @Override
     protected Settings nodeSettings(int nodeOrdinal) {
-        return Settings.builder().put(super.nodeSettings(nodeOrdinal)).put(settings).build();
+        Settings.Builder builder = Settings.builder();
+        if (REMOTE_CLUSTER_STATE_ENABLED_SETTING.get(settings)) {
+            if (repositoryPath == null) {
+                repositoryPath = randomRepoPath().toAbsolutePath();
+            }
+            builder.put(remoteStoreClusterSettings(REPOSITORY_NAME, repositoryPath));
+        }
+        return builder.put(super.nodeSettings(nodeOrdinal)).put(settings).build();
     }
 
     @Override
